+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHNQ7D73PBX0A8BSQZS0WX0R"
+        "01KHP41GJS9CCC7Y7D22KX2HCE"
     )
     r"""The unique identifier of the data source"""