+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHNQ7D6RXSPTSPYMVFZ4AGBY"
+        "01KHP41GJDNMBY8VNK4TBB4NWZ"
     )
     r"""The unique identifier of the data source"""