+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHNQ7D72H3J8HVVAZK202CCB"
+        "01KHP41GJGZF3F3J54FGMDZ4T6"
     )
     r"""The unique identifier of the data source"""