+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHNQ7D76VHC1RP3RXAASKDY3"
+        "01KHP41GJVQZJ91GVJSJX6F20C"
     )
     r"""The unique identifier of the data source"""