+++ packages/orq-rc/RELEASES.md
@@ -5739,3 +5739,13 @@ Based on:
 - [python v4.4.0-rc.5] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.5] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.5 - packages/orq-rc
+
+## 2026-02-17 15:41:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.721.0-rc.0 (2.827.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.6] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.6] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.6 - packages/orq-rc